Fix build warning about maven-source-plugin missing version
I noticed that while running ```mvn package``` it would complain that
``` [WARNING] 'build.plugins.plugin.version'
for org.apache.maven.plugins:maven-source-plugin is missing.```. This
commit aims to reolve that warning by specifying the most recent version
of the maven-source-plugin in pom.xml. According to
https://mvnrepository.com/artifact/org.apache.maven.plugins/maven-source-plugin,
the latest version of the maven-source-plugin is 3.2.1