Fix JsonTest#customModule on Windows
By using the absolute path, this test creates a path with invalid
characters in Windows.
One way to solve this is to use the URI instead of the absolute path
in Windows.

Note: Using the URI should be correct on any platform, but I wanted to
limit the scope of this patch as much as possible.